+++ src/main/java/pl/betoncraft/betonquest/compatibility/protocollib/hider/EntityHider.java
@@ -1,4 +1,4 @@
-package pl.betoncraft.betonquest.compatibility.protocollib;
+package pl.betoncraft.betonquest.compatibility.protocollib.hider;
 
 import com.comphenix.protocol.PacketType;
 import com.comphenix.protocol.ProtocolLibrary;
@@ -19,23 +19,20 @@ import org.bukkit.event.entity.EntityDeathEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.world.ChunkUnloadEvent;
 import org.bukkit.plugin.Plugin;
-import pl.betoncraft.betonquest.utils.LogUtils;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.Arrays;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.logging.Level;
 
-
-@SuppressWarnings({"PMD.TooManyMethods", "PMD.CommentRequired"})
+/**
+ * EntityHider From: https://gist.github.com/aadnk/5871793
+ */
+@SuppressWarnings("PMD.CommentRequired")
 public class EntityHider implements Listener {
-
     // Packets that update remote player entities
     private static final PacketType[] ENTITY_PACKETS = {
             PacketType.Play.Server.ENTITY_EQUIPMENT,
+            PacketType.Play.Server.BED,
             PacketType.Play.Server.ANIMATION,
             PacketType.Play.Server.NAMED_ENTITY_SPAWN,
             PacketType.Play.Server.COLLECT,
@@ -46,6 +43,7 @@ public class EntityHider implements Listener {
             PacketType.Play.Server.ENTITY_VELOCITY,
             PacketType.Play.Server.REL_ENTITY_MOVE,
             PacketType.Play.Server.ENTITY_LOOK,
+            PacketType.Play.Server.ENTITY_MOVE_LOOK,
             PacketType.Play.Server.ENTITY_TELEPORT,
             PacketType.Play.Server.ENTITY_HEAD_ROTATION,
             PacketType.Play.Server.ENTITY_STATUS,
@@ -93,16 +91,16 @@ public class EntityHider implements Listener {
      * @param observer - the observer player.
      * @param entityID - ID of the entity that will be hidden or made visible.
      * @param visible  - TRUE if the entity should be made visible, FALSE if not.
-     * @return TRUE if the entity was visible before this method call, FALSE
-     * otherwise.
+     * @return TRUE if the entity was visible before this method call, FALSE otherwise.
      */
-    protected boolean updateVisibility(final Player observer, final int entityID, final boolean visible) {
+    @SuppressWarnings("PMD.LinguisticNaming")
+    protected boolean setVisibility(final Player observer, final int entityID, final boolean visible) {
         switch (policy) {
             case BLACKLIST:
                 // Non-membership means they are visible
-                return !updateMembership(observer, entityID, !visible);
+                return !setMembership(observer, entityID, !visible);
             case WHITELIST:
-                return updateMembership(observer, entityID, visible);
+                return setMembership(observer, entityID, visible);
             default:
                 throw new IllegalArgumentException("Unknown policy: " + policy);
         }
@@ -117,7 +115,8 @@ public class EntityHider implements Listener {
      * @return TRUE if they already were present, FALSE otherwise.
      */
     // Helper method
-    protected boolean updateMembership(final Player observer, final int entityID, final boolean member) {
+    @SuppressWarnings("PMD.LinguisticNaming")
+    protected boolean setMembership(final Player observer, final int entityID, final boolean member) {
         if (member) {
             return observerEntityMap.put(observer.getEntityId(), entityID, true) != null;
         } else {
@@ -133,28 +132,21 @@ public class EntityHider implements Listener {
      * @return TRUE if they are present, FALSE otherwise.
      */
     protected boolean getMembership(final Player observer, final int entityID) {
-        try {
-            return observerEntityMap.contains(observer.getEntityId(), entityID);
-        } catch (UnsupportedOperationException e) {
-            LogUtils.getLogger().log(Level.WARNING, "Could not execute an operation: " + e.getMessage());
-            LogUtils.logThrowable(e);
-            return policy == Policy.WHITELIST;
-        }
+        return observerEntityMap.contains(observer.getEntityId(), entityID);
     }
 
     /**
      * Determine if a given entity is visible for a particular observer.
      *
      * @param observer - the observer player.
-     * @param entityID - ID of the entity that we are testing for visibility.
+     * @param entityID -  ID of the entity that we are testing for visibility.
      * @return TRUE if the entity is visible, FALSE otherwise.
      */
     protected boolean isVisible(final Player observer, final int entityID) {
-        // If we are using a whitelist, presence means visibility - if not, the opposite
-        // is the case
+        // If we are using a whitelist, presence means visibility - if not, the opposite is the case
         final boolean presence = getMembership(observer, entityID);
 
-        return (policy == Policy.WHITELIST) == presence;
+        return policy == Policy.WHITELIST ? presence : !presence;
     }
 
     /**
@@ -166,11 +158,8 @@ public class EntityHider implements Listener {
     protected void removeEntity(final Entity entity, final boolean destroyed) {
         final int entityID = entity.getEntityId();
 
-        final List<Map.Entry<Integer, Map<Integer, Boolean>>> list = new CopyOnWriteArrayList<>(observerEntityMap.rowMap().entrySet());
-        final Iterator<Map.Entry<Integer, Map<Integer, Boolean>>> iterator = list.iterator();
-
-        while (iterator.hasNext()) {
-            iterator.next().getValue().remove(entityID);
+        for (final Map<Integer, Boolean> maps : observerEntityMap.rowMap().values()) {
+            maps.remove(entityID);
         }
     }
 
@@ -191,23 +180,22 @@ public class EntityHider implements Listener {
      */
     private Listener constructBukkit() {
         return new Listener() {
-
-            @EventHandler(ignoreCancelled = true)
             @SuppressFBWarnings("UMAC_UNCALLABLE_METHOD_OF_ANONYMOUS_CLASS")
+            @EventHandler
             public void onEntityDeath(final EntityDeathEvent event) {
                 removeEntity(event.getEntity(), true);
             }
 
-            @EventHandler(ignoreCancelled = true)
             @SuppressFBWarnings("UMAC_UNCALLABLE_METHOD_OF_ANONYMOUS_CLASS")
+            @EventHandler
             public void onChunkUnload(final ChunkUnloadEvent event) {
                 for (final Entity entity : event.getChunk().getEntities()) {
                     removeEntity(entity, false);
                 }
             }
 
-            @EventHandler(ignoreCancelled = true)
             @SuppressFBWarnings("UMAC_UNCALLABLE_METHOD_OF_ANONYMOUS_CLASS")
+            @EventHandler
             public void onPlayerQuit(final PlayerQuitEvent event) {
                 removePlayer(event.getPlayer());
             }
@@ -215,21 +203,19 @@ public class EntityHider implements Listener {
     }
 
     /**
-     * Construct the packet listener that will be used to intercept every
-     * entity-related packet.
+     * Construct the packet listener that will be used to intercept every entity-related packet.
      *
      * @param plugin - the parent plugin.
      * @return The packet listener.
      */
     private PacketAdapter constructProtocol(final Plugin plugin) {
         return new PacketAdapter(plugin, ENTITY_PACKETS) {
-
             @Override
             public void onPacketSending(final PacketEvent event) {
-                final int index = event.getPacketType() == PacketType.Play.Server.COMBAT_EVENT ? 1 : 0;
+                final int entityID = event.getPacket().getIntegers().read(0);
 
-                final Integer entityID = event.getPacket().getIntegers().readSafely(index);
-                if (entityID != null && !isVisible(event.getPlayer(), entityID)) {
+                // See if this packet should be cancelled
+                if (!isVisible(event.getPlayer(), entityID)) {
                     event.setCancelled(true);
                 }
             }
@@ -239,8 +225,7 @@ public class EntityHider implements Listener {
     /**
      * Toggle the visibility status of an entity for a player.
      * <p>
-     * If the entity is visible, it will be hidden. If it is hidden, it will become
-     * visible.
+     * If the entity is visible, it will be hidden. If it is hidden, it will become visible.
      *
      * @param observer - the player observer.
      * @param entity   - the entity to toggle.
@@ -263,7 +248,7 @@ public class EntityHider implements Listener {
      */
     public final boolean showEntity(final Player observer, final Entity entity) {
         validate(observer, entity);
-        final boolean hiddenBefore = !updateVisibility(observer, entity.getEntityId(), true);
+        final boolean hiddenBefore = !setVisibility(observer, entity.getEntityId(), true);
 
         // Resend packets
         if (manager != null && hiddenBefore) {
@@ -279,9 +264,10 @@ public class EntityHider implements Listener {
      * @param entity   - the entity to hide.
      * @return TRUE if the entity was previously visible, FALSE otherwise.
      */
+    @SuppressWarnings("PMD.AvoidThrowingRawExceptionTypes")
     public final boolean hideEntity(final Player observer, final Entity entity) {
         validate(observer, entity);
-        final boolean visibleBefore = updateVisibility(observer, entity.getEntityId(), false);
+        final boolean visibleBefore = setVisibility(observer, entity.getEntityId(), false);
 
         if (visibleBefore) {
             final PacketContainer destroyEntity = new PacketContainer(PacketType.Play.Server.ENTITY_DESTROY);
@@ -290,8 +276,8 @@ public class EntityHider implements Listener {
             // Make the entity disappear
             try {
                 manager.sendServerPacket(observer, destroyEntity);
-            } catch (InvocationTargetException e) {
-                LogUtils.getLogger().log(Level.WARNING, "Cannot send server packet.", e);
+            } catch (final InvocationTargetException e) {
+                throw new RuntimeException("Cannot send server packet.", e);
             }
         }
         return visibleBefore;
@@ -300,14 +286,13 @@ public class EntityHider implements Listener {
     /**
      * Determine if the given entity has been hidden from an observer.
      * <p>
-     * Note that the entity may very well be occluded or out of range from the
-     * perspective of the observer. This method simply checks if an entity has been
-     * completely hidden for that observer.
+     * Note that the entity may very well be occluded or out of range from the perspective
+     * of the observer. This method simply checks if an entity has been completely hidden
+     * for that observer.
      *
      * @param observer - the observer.
      * @param entity   - the entity that may be hidden.
-     * @return TRUE if the player may see the entity, FALSE if the entity has been
-     * hidden.
+     * @return TRUE if the player may see the entity, FALSE if the entity has been hidden.
      */
     public final boolean canSee(final Player observer, final Entity entity) {
         validate(observer, entity);
@@ -340,11 +325,12 @@ public class EntityHider implements Listener {
 
     /**
      * The current entity visibility policy.
+     *
+     * @author Kristian
      */
     public enum Policy {
         /**
-         * All entities are invisible by default. Only entities specifically made
-         * visible may be seen.
+         * All entities are invisible by default. Only entities specifically made visible may be seen.
          */
         WHITELIST,