+++ src/main/java/pl/betoncraft/betonquest/compatibility/protocollib/hider/UpdateVisibilityNowEvent.java
@@ -1,4 +1,4 @@
-package pl.betoncraft.betonquest.compatibility.protocollib;
+package pl.betoncraft.betonquest.compatibility.protocollib.hider;
 
 import pl.betoncraft.betonquest.Instruction;
 import pl.betoncraft.betonquest.api.QuestEvent;
@@ -8,6 +8,7 @@ import pl.betoncraft.betonquest.utils.PlayerConverter;
 
 @SuppressWarnings("PMD.CommentRequired")
 public class UpdateVisibilityNowEvent extends QuestEvent {
+    private static PlayerHider hider;
 
     public UpdateVisibilityNowEvent(final Instruction instruction) throws InstructionParseException {
         super(instruction, true);
@@ -16,6 +17,13 @@ public class UpdateVisibilityNowEvent extends QuestEvent {
     @Override
     protected Void execute(final String playerID) throws QuestRuntimeException {
         NPCHider.getInstance().applyVisibility(PlayerConverter.getPlayer(playerID));
+        if(hider != null) {
+            hider.updateVisibility();
+        }
         return null;
+    }
+
+    public static void setHider(final PlayerHider playerHider) {
+        hider = playerHider;
     }
 }